+++ scripts/constants.sh
@@ -4,20 +4,26 @@
 GOPATH="$(go env GOPATH)"
 
 # Set binary location
-binary_path=${CAMINOETHVM_BINARY_PATH:-"$GOPATH/src/github.com/chain4travel/caminogo/build/plugins/evm"}
+binary_path=${CAMINOETHVM_BINARY_PATH:-"$GOPATH/src/github.com/chain4travel/camino-node/build/plugins/evm"}
 
 # Avalabs docker hub
-dockerhub_repo="avaplatform/caminogo"
+dockerhub_repo="c4tplatform/caminogo"
 
-# Current branch
-current_branch=${CURRENT_BRANCH:-$(git describe --tags --exact-match 2> /dev/null || git symbolic-ref -q --short HEAD || git rev-parse --short HEAD)}
-echo "Using branch: ${current_branch}"
+# Current tag
+caminoethvm_tag=${CURRENT_BRANCH:-$(git describe --tags --exact-match 2> /dev/null || git symbolic-ref -q --short HEAD || git rev-parse --short HEAD)}
+echo "Using tag: ${caminoethvm_tag}"
 
 # Image build id
+caminoethvm_commit=${CAMINOETHVM_COMMIT:-$( git rev-list -1 HEAD )}
 # Use an abbreviated version of the full commit to tag the image.
+caminoethvm_short_commit="${caminoethvm_commit::8}"
+caminoethvm_tag=${CAMINOETHVM_TAG:-$( git describe --tags --dirty )}
 
-# WARNING: this will use the most recent commit even if there are un-committed changes present
-caminoethvm_commit="$(git --git-dir="$CAMINOETHVM_PATH/.git" rev-parse HEAD)"
-caminoethvm_commit_id="${caminoethvm_commit::8}"
+# caminogo version
+module=$(grep caminogo $CAMINOETHVM_PATH/go.mod)
+# trim leading
+module="${module#"${module%%[![:space:]]*}"}"
+t=(${module//\ / })
+caminogo_tag=${t[-1]}
 
-build_image_id=${BUILD_IMAGE_ID:-"$camino_version-$caminoethvm_commit_id"}
+build_image_id=${BUILD_IMAGE_ID:-"$caminogo_tag-$caminoethvm_short_commit"}