+++ scripts/versions.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-# Set up the versions to be used
-caminoethvm_version=${CAMINOETHVM_VERSION:-'v0.0.1'}
-# Don't export them as they're used in the context of other calls
-camino_version=${CAMINO_VERSION:-'v0.0.2'}